Decission Tree Algorithm implementation
A decision tree is a supervised machine learning algorithm that is used for classification and regression tasks. 
It is a flowchart-like structure in which each internal node represents a feature or attribute, each branch represents a decision rule, and each leaf node represents an outcome or a class label.
The process of building a decision tree involves selecting the best feature at each step to split the data. 
The selection is usually based on metrics such as information gain, Gini impurity, or entropy, which quantify the effectiveness of a particular feature in reducing uncertainty or improving the classification accuracy. 
The algorithm continues recursively until a stopping criterion is met, such as reaching a maximum depth or when all the instances in a leaf node belong to the same class.
Decision trees have several advantages, including their interpretability and ability to handle both categorical and numerical data.