Decision Tree Classifier Iris_Breast_Cancer
A decision tree classifier is a machine learning algorithm used for classification tasks. It builds a tree-like model of decisions and their possible consequences based on a given set of features. 
The tree consists of internal nodes that represent features and branches that represent the decision rules based on those features. At the leaf nodes, the decision tree assigns a class label or a probability distribution of class labels to each instance.

The decision tree classifier operates by recursively partitioning the data based on the values of the features, aiming to create homogeneous subsets of instances within each partition. The splitting process is driven by metrics such as information gain, Gini index, or entropy, which measure the purity or impurity of the subsets. 
The goal is to maximize the homogeneity of the subsets while minimizing the impurity.

To classify a new instance, it traverses the decision tree by evaluating the feature conditions at each internal node and following the appropriate branch until reaching a leaf node. The class label assigned at the leaf node is then used as the predicted class for the instance.

The decision tree classifier is interpretable, as the resulting tree structure can be visualized and understood by humans. It is also flexible, allowing for nonlinear relationships between features and the target variable. However, decision trees are prone to overfitting, especially when the tree becomes too deep and complex. To mitigate this, techniques such as pruning, setting minimum sample requirements for splitting, or using ensemble methods like random forests can be employed.