KNN Classifier and Regression
K-Nearest Neighbors (KNN) classifier is a simple machine learning algorithm that classifies data points based on their proximity to other data points in a feature space, where the class of the nearest neighbors determines the class of the input point. KNN regressor, on the other hand, predicts a continuous value for a given data point based on the average of the values of its nearest neighbors, making it useful for solving regression problems.